# apps/vc-api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/vc-api/credentials/issue": {
      "post": {
        "operationId": "VcApiController_issueCredential",
        "summary": "",
        "description": "Issues a credential and returns it in the response body. Conforms to https://w3c-ccg.github.io/vc-api/issuer.html",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/IssueCredentialDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/VerifiableCredentialDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalServerErrorResponseDto" }
              }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/credentials/verify": {
      "post": {
        "operationId": "VcApiController_verifyCredential",
        "summary": "",
        "description": "Verify a credential. Conforms to https://w3c-ccg.github.io/vc-api/#verify-credential",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/VerifyCredentialDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "Verifiable Credential successfully verified",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/VerificationResultDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    { "$ref": "#/components/schemas/VerificationResultDto" },
                    { "$ref": "#/components/schemas/BadRequestErrorResponseDto" }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalServerErrorResponseDto" }
              }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/presentations/prove/authentication": {
      "post": {
        "operationId": "VcApiController_proveAuthenticationPresentation",
        "summary": "",
        "description": "Issue a DIDAuth presentation that authenticates a DID.\nNot a part of VC-API? Maybe there is a DID Auth spec though?\nA NON-STANDARD endpoint currently.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthenticateDto" } } }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/VerifiablePresentationDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalServerErrorResponseDto" }
              }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/presentations/from": {
      "post": {
        "operationId": "VcApiController_presentationFrom",
        "summary": "",
        "description": "Creates a Presentation without Proof by passing in the Presentation Definition, selected Verifiable Credentials (TODO: and an optional holder (DID)).\nThe presentation contains the [presentation submission](https://identity.foundation/presentation-exchange/#presentation-submission) data that the verifier can use.",
        "parameters": [],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PresentationDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalServerErrorResponseDto" }
              }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/presentations/prove": {
      "post": {
        "operationId": "VcApiController_provePresentation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ProvePresentationDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/VerifiablePresentationDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalServerErrorResponseDto" }
              }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/presentations/verify": {
      "post": {
        "operationId": "VcApiController_verifyPresentation",
        "summary": "",
        "description": "Verify a presentation. Conforms to https://w3c-ccg.github.io/vc-api/#verify-presentation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/VerifyPresentationDto" } }
          }
        },
        "responses": {
          "200": {
            "description": "Verifiable Presentation successfully verified!",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/VerificationResultDto" } }
            }
          },
          "400": { "description": "Invalid or malformed input" },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalServerErrorResponseDto" }
              }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/exchanges": {
      "post": {
        "operationId": "VcApiController_createExchange",
        "summary": "",
        "description": "Allows the creation of a new exchange by providing the credential query and interaction endpoints\nA NON-STANDARD endpoint currently.\n\nSimilar to https://gataca-io.github.io/vui-core/#/Presentations/post_api_v2_presentations",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ExchangeDefinitionDto" } }
          }
        },
        "responses": {
          "201": { "description": "" },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ConflictErrorResponseDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalServerErrorResponseDto" }
              }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/exchanges/{exchangeId}": {
      "post": {
        "operationId": "VcApiController_initiateExchange",
        "summary": "",
        "description": "Initiates an exchange of information.\nhttps://w3c-ccg.github.io/vc-api/#initiate-exchange",
        "parameters": [
          { "name": "exchangeId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ExchangeResponseDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorResponseDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalServerErrorResponseDto" }
              }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
...
    "/v1/vc-api/exchanges/{exchangeId}/{transactionId}/review": {
      "post": {
        "operationId": "VcApiController_addSubmissionReview",
        "summary": "",
        "description": "Update a transaction review\nA NON-STANDARD endpoint currently.\nSimilar to https://github.com/energywebfoundation/ssi-hub/blob/8b860e7cdae4e1b1aa75afeab8b9df7ab26befbb/src/modules/claim/claim.controller.ts#L80",
        "parameters": [
          { "name": "exchangeId", "required": true, "in": "path", "schema": { "type": "string" } },
          { "name": "transactionId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/SubmissionReviewDto" } }
          }
        },
        "responses": {
          "201": { "description": "" },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorResponseDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalServerErrorResponseDto" }
              }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/workflows": {
      "post": {
        "operationId": "VcApiController_createWorkflow",
        "summary": "",
        "description": "Creates a new workflow and returns its information in the response body.\nSee https://w3c-ccg.github.io/vc-api/#create-workflow",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateWorkflowRequestDto" } }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CreateWorkflowSuccessDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ConflictErrorResponseDto" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalServerErrorResponseDto" }
              }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
    "/v1/vc-api/workflows/{localWorkflowId}": {
      "get": {
        "operationId": "VcApiController_getWorkflow",
        "summary": "",
        "description": "Gets the configuration of an existing workflow and returns it in the response body.\nSee https://w3c-ccg.github.io/vc-api/#get-workflow-configuration",
        "parameters": [
          { "name": "localWorkflowId", "required": true, "in": "path", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CreateWorkflowSuccessDto" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/BadRequestErrorResponseDto" } }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundException" } }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InternalServerErrorResponseDto" }
              }
            }
          }
        },
        "tags": ["vc-api"]
      }
    },
...
  },
  "info": { "title": "VC-API", "description": "Sample VC-API", "version": "0.1", "contact": {} },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "BadRequestErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "example": 400 },
          "message": {
            "oneOf": [
              { "type": "array", "items": { "type": "string" }, "example": ["error 1", "error 2"] },
              { "type": "string", "example": "error" }
            ]
          },
          "error": { "type": "string", "example": "Bad Request" }
        },
        "required": ["statusCode", "message", "error"]
      },
...
      "IssueCredentialDto": {
        "type": "object",
        "properties": {
          "credential": { "$ref": "#/components/schemas/CredentialDto" },
          "options": { "$ref": "#/components/schemas/IssueOptionsDto" }
        },
        "required": ["credential"]
      },
      "VerifiableCredentialDto": {
        "type": "object",
        "properties": {
          "@context": {
            "type": "array",
            "description": "The JSON-LD context of the credential.",
            "items": { "oneOf": [{ "type": "string" }, { "type": "object" }] }
          },
          "id": { "type": "string", "description": "The ID of the credential." },
          "type": {
            "description": "The JSON-LD type of the credential.",
            "type": "array",
            "items": { "type": "string" }
          },
          "issuer": {
            "description": "A JSON-LD Verifiable Credential Issuer.",
            "oneOf": [{ "$ref": "#/components/schemas/IssuerDto" }, { "type": "string" }]
          },
          "issuanceDate": {
            "type": "string",
            "description": "The issuanceDate",
            "example": "2022-09-21T11:49:03.205Z"
          },
          "expirationDate": {
            "type": "string",
            "description": "The expirationDate",
            "example": "2023-09-21T11:49:03.205Z"
          },
          "credentialSubject": { "type": "object", "description": "The subject" },
          "proof": { "type": "object", "description": "A JSON-LD Linked Data proof." }
        },
        "required": ["@context", "type", "issuer", "issuanceDate", "credentialSubject", "proof"]
      },
...
      "AuthenticateDto": {
        "type": "object",
        "properties": {
          "did": { "type": "string" },
          "options": { "$ref": "#/components/schemas/ProvePresentationOptionsDto" }
        },
        "required": ["did", "options"]
      },
...
    }
  }
}